requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
redis>=5.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
rich>=13.0.0
typer>=0.12.0
